tests/wpt/web-platform-tests/2dcontext/conformance-requirements/2d.coordinatespace-manual.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.coordinatespace</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/2dcontext/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.coordinatespace</h1>
<p class="desc">Coordinate space goes from top-left to bottom-right</p>

<p class="notes">This should not be upside down.
<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="2d.coordinatespace.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("Coordinate space goes from top-left to bottom-right");
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#00f';
ctx.fillRect(0, 0, 100, 50);
ctx.fillStyle = '#0ff';
ctx.fillRect(0, 0, 50, 25);
_assertPixel(canvas, 25,12, 0,255,255,255, "25,12", "0,255,255,255");
_assertPixel(canvas, 75,12, 0,0,255,255, "75,12", "0,0,255,255");
_assertPixel(canvas, 25,37, 0,0,255,255, "25,37", "0,0,255,255");
_assertPixel(canvas, 75,37, 0,0,255,255, "75,37", "0,0,255,255");


});
</script>
